+++ lib/forge_sdk/util.ex
@@ -132,8 +132,6 @@ defmodule ForgeSdk.Util do
 
   def datetime_to_proto(dt), do: Google.Protobuf.Timestamp.new(seconds: DateTime.to_unix(dt))
 
-  def proto_to_datetime(%{seconds: seconds}), do: DateTime.from_unix!(seconds)
-
   def token_to_unit(n, name \\ "") do
     conn = ForgeSdk.get_conn(name)
     ForgeAbi.token_to_unit(n, conn.decimal)